Install puppet modules in right order
Associative arrays in bash are traversed in random order,
so it means that even pinning module versions and declaring
them prior to modules that depend on them, this doesn't mean
that pinned version of the module is going to be installed.

If a module that has a dependency with another module installs a
higher version, then we try to install a lower pinned version
of the dependent module, this will result in failure because higher
version of module is already installed.

Declaring modules in the correct install order in an indexed array
and traversing that instead of the associative one will solve the
problem as modules with pinned versions can be installed first
and forced to that.

Change-Id: I980539287e5ffd86be30f5e2fe2dcc991f7e8b9e