Add mysql-proxy to enable read-only access to a db
This commit adds a mysql_proxy module which will setup a read-only
proxy to a mysql db. The module also adds a dir for adding lua
scripts which is how you can define custom rules for the proxy. While
not used now, this will enable future usage if we determine we need
more complex rules in the proxy configuration.

Change-Id: I478baca354354347fe50074a8e3b9f66ca890d55