Switch hpcloud-b5 to devstack-trusty dib image
Nodepool supports dib now. We have dib elements to build devstack
images. Now we can put the two together and take over the world with dib
and glance. Lets do this, starting small and using a single nodepool
provider. A future change should switch all devstack-trusty nodes to dib
images and so on.

Change-Id: I179df9117be538c3ab44cce3a3cd579e03e21e90