Add new experimental jobs to test dib based nodes
We now have new dib based slave nodes, the next step in moving to dib is
to test that these images are viable for testing. Run experimental jobs
that are otherwise identical to two of our existing gate jobs (one with
neutron the other without) on these new dib based slaves.

Change-Id: If69f18a61529fe886a836c6c284103babbcfc2cb