Add centos7 image to nodepool d-i-b config
Change Ida38cba0a303ab77432d5d354f1952e00ebfa00e adds support to
tools/build-image.sh for building test-images with these elements and
release.  This adds support for nodepool to build the image and deploy
on hpcloud.

Note this retains the original centos7 nodes using the RAX provided
image, which are currently only used in an experimental devstack job.
These will be removed for this when stable.

Change-Id: Ic1848b98625603bf37c0e3dfaf1d0ef6f2c6ac74