Upload custom configuration for js-draft build.
With CORS support in the API, and custom configuration loading in the
webclient, we can now let our js-draft build connect to a live API and
function like a regular client. This change makes the assumption that
any javascript project will include static docs-draft assets in the
$PROJECT_ROOT/test/draft directory, which will be uploaded as-is to the
same destination as the draft build.

Change-Id: If41a1a12b14c17c62baf4ce9473907f7a374ccc3