Removes zuul/layout-dev.yaml
The file zuul/layout-dev.yaml was moved to the project-config repo.
This copy of the file in config should have been deleted with the
others in commit 1e7874e1eb0c15adc06d0bd10da5a90a26901309 and it
wasn't. There have been no updates to this file since August 5,
2014. This patch deletes zuul/layout-dev.yaml.

Change-Id: I4d6729d289b00b3ce05aa98f27d275b5093b9f00