Reload the gerrit header with its dependencies
When gerrit loads a header it adds the md5sum of the resources
linked from within the header to the uri (eg
static/hideci.js?e=1231d06b0e69a2dc7cf953b47f3b9d853b9). This is
then evaluated when the static content is requested from the gerrit-
server. As such when the md5 of one of the resources changes we need
to reload the header for new sums to be calculated.

Change-Id: I330f123d3eff6109539d9f06f80c8212a33781e1