Deploy vhost template for nodepool.o.o logs
In Ie0b269835ebb8effbac0285b782d8add7b47db32 I didn't consider how
difficult it is to get puppet to put in AllowOverride directives to
allow .htaccess to work.  The puppetlabs-apache version we use does
not support "override" as an argument, so you can't easily add this to
allow apache to read htaccess.  Upgrading is a big task because we are
so far behind (see I6fa5f3275a84ec4480169de562d1d4a656111814).

So deploy a full template for nodepool vhost that includes the config
options we need.

Additionally, the existing mimetype isn't set for the log-files

---
 $ curl -I http://nodepool.openstack.org/image.log.2014-10-14
 HTTP/1.1 200 OK
 Date: Fri, 17 Oct 2014 00:02:39 GMT
 Server: Apache/2.2.22 (Ubuntu)
 Last-Modified: Tue, 14 Oct 2014 23:59:56 GMT
 ETag: "f89bb-d458315-5056acfe33700"
 Accept-Ranges: bytes
 Content-Length: 222659349
---

I'm pretty sure this means it defaults to text/plain, but this makes
it explicit.  This should also matches on the timestampped log files
per the rules of multiple extensions [1]

[1] http://httpd.apache.org/docs/2.2/mod/mod_mime.html#multipleext

Change-Id: I7fa2603f4160b06af71a515e655d4a705fa0c768