Bail out in run_all.sh runs when failures happen
Previously the run_all.sh script would run puppet even if git is not
properly updated or if our puppet modules fail to install. Now set -e in
the script so that any failure to update git or install puppet modules
causes run_all.sh to bail out early.

This is important to ensure that we get consistent and expected results
from puppet when it runs.

Change-Id: Icb3fb2a97d11675762b49c57978b08115bfcbc04