Experimental: Livegrep implementation
This is an experimental livegrep implementation. It uses puppet
vcsrepo to watch the upstream git repository (currently mine,
just for testing). When that repo changes, it fires events that
rebuild docker containers. This eventually bounces the docker
containers or something.

I haven't figured out how to roll updates though the docker
containers. There are three solutions I can see to that:

1) Just kill em, restart em.

2) Install kubernetes and etcd.

3) Use nova::docker

4) Don't use puppet to manage worker pools? How does storyboard
do this?

Change-Id: I381d41baa6526836e9d00d34cf2397bb0dfb41f1