fix hideci for new change screen
The previous hideci enhancements and refactor was not tested on the
new change screen, which sadly removes all the semantic markup around
the displayed changes (boo gerrit).

Do the base parsing based on the looping over the p tags as was done
in the initial version instead of matching divs by semantic tags. A
few of the parsing functions needs to be adapted to know if they are
on old vs. new change screen (like finding the date).

Spot tested locally with new and old change screen, seems to be
working.

Change-Id: Ieadfb78fa4d152d12cb64fc69d808fa510d28abd