Loosen name restrictions for dedicated 3rd-parties
* doc/source/third_party.rst: Clarify the third-party account
request recommendations to accommodate the "Community CI" use case,
where a specific project's community collaboratively maintains an
external testing system solely for reporting on proposed changes to
its own projects. The "Organization Subsystem CI" name pattern is
not relevant in some of these cases, and it may be unavoidable that
the account mentions the name of the project it's testing in these
sorts of situations.

Change-Id: I70f1a23f32ae6ee148753e00c167b79a7197fc93