Switch to installing subunit2sql from git instead of pip
This commit switches to using git to install subunit2sql from releases
instead of using the Pip class. The problem with the pip class is that
subscribing to it generates an event on each puppet refresh, which
causes the db-manage utility to run a migration update every 15 min,
which is far from ideal.

Since we only want to run the updates on each pushed release of
subunit2sql and we don't want to CD the git repo, the most recent tag
is specified. This will be required to be manually updated for each
subunit2sql release, but this is a tradeoff between not being able to
use the pip class to trigger migrations and CDing subunit2sql which
isn't recommended by the maintainer of the package.

Change-Id: Ib77d68f19edbf48a8d8cf4e757dbfb6463f1c6bb