Set dib vars to override tmp and cache locations
We need to override the locations of tmp and DIB cache to take advantage
of ephemeral disks on cloud nodes. DIB requires many gigabytes of
storage in our use of the tool and many cloud VMs simply do not provide
that on the root filesystem.

In our specific case we mount the ephemeral drive on /opt so we use
/opt/dib_cache and /opt/dib_tmp as the new locations for these
directories when running DIB.

Change-Id: Id126d3e63a518886c8047684ae14e891f6fdad99