Add a debootstrap-based Ubuntu element
As a first step towards having our own base images, make a new
root image element that is based on debootstrap and not on the
ubuntu cloud images. This means we can avoid things like cloud-init or
the giant host of other stuff we don't want. Will follow this up with
an element that adds our basic keys, so that our node launching can
also skip using the keypair extension (which is what we need cloud-init
for)

Change-Id: I8d59d77394a94b5ef915ceb88b7d18d1c358960a