Set modulepath to the first element of puppet's modulepath.
This duplicates the behavior of puppet module install for the git
clone operations.

Puppet module install will use the first element of the modulepath.
This makes the git clone operations go into the same directory as
the modules installed via the puppet module tool.

This helps us downstream because we are moving the directory that
holds these modules out of /etc/ and into /opt/.

Change-Id: I0c7833a6d9a79fa613efc9daebee916f3192306a