Make localuser gid and uid required params
Before we had set gid and uid to unset by default. Which if not
overridden will have puppet try to set the gid and uid values to
'unset'. This should be an error as 'unset' is not a number. We don't
have this error because we always override these values when creating
localusers. But for completeness and better error messages remove the
default values so that puppet requires users to have proper gid and uid
values.

Change-Id: I14667912b9239ea1c56123c18b0c13834f9c3cb6