Added manifest for storyboard-dev.openstack.org
This manifest adds a configuration for a storyboard dev machine. It
will require additional infra assets in hiera, including a new SSL cert,
credentials, and a new trove database (MySQL 5.6).

This patch is dependent on https://review.openstack.org/#/c/144367,
and will need to be rechecked once that patch lands.

Change-Id: Iaec515c2a8512c6b3a1ce0ec6dce0ce81fc4952f