Set qemu-img qcow2 compat level to 0.10
New qemu-img on trusty defaults to compat level 1.1 when building qcow2
images. This is a problem when clouds have older qemu-img versions that
only support 1.0 and older. Override the compat level and set it to 0.10
as suggested in qemu-img's manpage to ensure compatibility with clouds
when building nodepool disk image builder images.

Change-Id: I3edd209e9562a99e5efaea5fa59872a8e3460ef0