Enable websockets transport on Etherpad
This can enhance the user experience, as most modern web browsers
are capable of using Websockets.
Due to SocketIO fallback mechanism, if Websocket is not available
the next transport in the list will be used, so this change would
still work like before where Websockets is not possible.
Please note that Apache must be able to pass thru websockets to
the Etherpad nodejs server, the module mod_proxy_wstunnel is
needed for that, but it's only available as a package for
Apache 2.4.
I will create another change for etherpad-dev that adds
that module and sync up with infra root members to upgrade the server
to Trusty (it comes with Apache 2.4, Precise still comes with 2.2).

Change-Id: I22dbfc971438e726ed5796f714624bb1216a15d3