Only run one DIB image "online"
DIB has cache contention issues when multiple image builds start
together because they all want to write to and update the cache. Avoid
this by running all but one image build offline. This leaves us with one
build (the first one) that is responsible for updating the cache.

Note this is a bad hack and we should more properly make DIB deal with
this problem. Perhaps run all builds offline and have some other out of
band process update the cache for us.

Change-Id: If4831ef5c8c2e8c97e6ab718f2ace13041a1edf6