Make project_config branch variable
When setting up 3-dparty CI based on OpenStack CI manifests we also
need to keep track of the lastest changes introduced to the upstream.
To do this we need to keep 'master' branch in sync with upsream, keeping
our patches in a separate branch. Thus we must be able to install the CI
from that custom branch.

This commit makes it possible to use a custom branch for project_config,
with 'master' brach as the default.

Change-Id: I1d4d3fd4cb4349c47d84b71497ac248a26ed73d1