Split image build logs for nodepool
Currently all the image build logs are squished together into
image.log, which often grows to hundreds of megabytes and must be
painfully manually extracted.  It makes it far too easy to miss broken
builds.

My first inclination was to get nodepool to split these log files
internally.

The fundamental issue with that is that in python's logging module, a
FileHandler based handler has a 1-to-1 relationship with an output
file (this includes classes built on that like RotatingFileHandler,
etc).  Many different loggers can send their messages to a single
handler, but it will only open one stream and write out one file.

This would be OK if we instantiated a new handler for every image
build; e.g. each build does
logging.getLogger('nodepool.image.build.foo.bar') and we instantiate a
new handler outputting to the relevant "foo.bar" log file for this
logger.

This is certainly possible, but we have the issue of configuring this.
Currently, logging setup is done in a separate logging config read by
logging.config.fileConfig().  This sets up the loggers, handlers and
filters at program load and goes away.

If we want to instantiate new loggers dynamically at image build time,
we need information on the handler class to use, arguments to the
class, formatters, etc.  So how could we get this:

 - parse the logging config ourselves and re-use the
   nodepool.image.build logger config but override the file-name
   (essentially re-writing logging.config.fileConfig() in nodepool)
 - embed this information within the nodepool configuration;
   duplicating the logging config file but in yaml and making
   image-build logging special
 - we could hard-code the handler and not make it configurable; which
   is a regression from current configurability; but we still need a
   path to put the file (do we re-parse the logging config again or
   add it to nodepool conf, rinse, repeat)

We could write our own custom handler that keeps multiple streams open
and directs log messages from different loggers to different files.
But we can't base that on the FileHandler classes, so they are special
and different, and we have to write our own rotation, etc, which we
will get wrong.

I did also consider doing some introspection on the
"nodepool.image.build" logger created by the config file, trying to
pick-out it's handler at runtime and using that to dynamically create
a new handler for sub-loggers.  That is rather insane when you start
thinking through the implications of people setting up multiple
handlers, etc.

Thus the best solution is to retain the status-quo of the logging
config file being the cannonical source of logging data and avoid
turning image logging into some special case.  We can simply create
the required loggers and handlers up front.

To do this, we add generate-nodepool-loggers.py which generates the
nodepool.logging.conf file based upon the providers and images in the
nodepool configuration file.

Nodepool already creates a "nodepool.image.build.provider.image"
logger for image builds (bootstrapServer() [1]).

The disadvantage of this method is that for newly added or modified
images, they will not have a specific logger/handler setup for them
until the config file is regenerated and nodepool reloaded.  In this
case, their messages will bubble-up to the generic image.log file just
as before.  Given the rate of change of images, I think this is a good
compromise.

[1] https://git.openstack.org/cgit/openstack-infra/nodepool/tree/nodepool/nodepool.py

Change-Id: I648b7fe0ef7446145bc1e0abc89f5e8ba9ffb9e7