Make sure Gerrit sites are using the proper libraries.
Running the gerrit 'init' command to create a review site will
automatically download and install bouncycastle and mysql libs
into the $SITE/lib folder. The versions of those libs can vary
depending on which version of Gerrit the command is run from.

For example:

Gerrit 2.8.x installs:
  mysql-connector-java-5.1.10.jar
  bprov-jdk16-1.44.jar
  bcpg-jdk16-1.44.jar

Gerrit 2.9.x installs:
  mysql-connector-java-5.1.21.jar
  bprov-jdk15on-1.49.jar
  bcpg-jdk15on-1.49.jar
  bbcpkix-jdk15on-1.49.jar

This change makes sure we are installing and using the
equivalent libs from the ubuntu package repo.

Notes:
[1] Gerrit 2.9.x requires libbbcpkix-java which is not available on
precise but is available on trusty.
[2] Have tested on review-dev.o.o that the updates libs from this change
will work for a Gerrit 2.8.x site as well

Change-Id: Id057782b9f350ea3f0728660cba056fc8e37c898