Expand ranges on iptables rules for floating IPs
We have a small set of iptables rules on our single use slaves that
enable ironic and heat functionality. We are shifting the floating IP
range from 172.24.4.0/24 to 172.24.5.0/24 and placing an overlapping
range of 172.24.4.0/23 to give compute nodes routes to the floating IPs
in multinode situations.

To accmodate these changes expand the existing rules to cover
172.24.4.0/23 instead of just 172.24.4.0/24.

Change-Id: I0b28c3607747c3939912ce4664627910f431dba6