Only start 1 subunit worker per node
This commit removes the 3 unneeded subunit worker processes from the
subunit_worker node type. Previously, 4 workers were launched on the
same node (similar to the logstash workers) but this caused issues
in production. (they consumed all the ram) After dropping down to one
process it was able to handle the load without issue and didn't
consume all the resources.

Change-Id: I5f7f85fc6f8bfb5d2fbec548b58c5d8a1bc0e628