A rebase button for all registered users
The nova project core reviewers have had this enabled for them for
several months without incident, and it's already something anyone
can do locally and push to a change. Also it's convenient when
people are editing commit messages via the WebUI and need to rebase
dependent changes after doing so.

Change-Id: Iefff52be4d12390be372429d6ebc48704c379630