Use nodepool 'generate-log-config' command to generate log config
Use nodepool's 'generate-log-config' command to generate the log
config for nodepool.  This relies on nodepool being deployed and being
able to read the nodepool configuration file.

Because this relies on the generated-from-template config file, this
has to run on the slave.  We regenerate the file with an exec, but
only update if it has changed -- this will facilitate reloading the
config in nodepoold without us constantly updating the file on it.

Change-Id: I90af348cea41e95fe0795bfef618073d4aabb1c2
Depends-On: I916f3dd9634ca7ac691bef98236f6cf549fb7059