Drive actions around repo creation from ansible
The only reason we need to order puppet runs is actually because we have
two sets of data that need to be actioned. puppet itself on git* and
gerrit is not order dependent. Also, rather than parameterizing the
input git ref to puppet, simplify and actually just drive it as an
action from ansible in the first place.

Change-Id: I9c2d52c88245d3ddce2eb8156bb684468fe38fb6