Start passing in data from puppetmaster hiera
As a step towards using puppet apply, pass in data to the remote hosts
that gets put into local hiera so that the hiera() calls in the puppet
manifests would have data to draw from. To make this clean, start using
in-tree group_vars to set defaults for the role parameters.

Change-Id: I5698355df0c13cd11fe5987787e65ee85a384256
Depends-On: I479390c97e3029415db258c1868a3facd1e6c00f