Install puppet modules on the right order
Currently modules are being installed without
predefined order. It supposes that modules with a
pinned version are givin errors, because modules
that bring a nested dependency, are installing a newer version.
Add the list of modules to be installed into a working
sequence, and install that on the right order.

Change-Id: Iaf91278458ab16caa511a557a5ac3cf4040a301e