Generate nodepool logging configuration
Generate a logging configuration template for nodepool that splits
image-build logs out into separate files.

A tool to generate and maintain this file based on the nodepool
configuration is added.

The logging configuration template this tool creates is added
alongside the main configuration.  This is passed to the deployment
puppet module via a parameter.

Change-Id: I4723c68e5c83b536560e0a8bd7b73e312b46fc02
Depends-on: Iec77fb62c353d09186140b23912caf02225a643b