Ensure latest on etherpad vcsrepo
Due to some implementation details in our vcsrepo puppet module in order
to get the specified revision you must ensure latest on the repo. This
poses a small problem where the repo is reset to that revision each
puppet run, but we can negate that problem by hosting the robots.txt
from apache directly which is where the depends on dependency in this
change comes from.

Change-Id: Ibcfb03cf9ff6f7686e5dbd69905d34009adf9157
Depends-On: Iccccf1ef194060490512e6550c22bdb9d3478ba8