Fix jquery setup on Gerrit server.
The libjs-query package sets up jquery with on trusty with these
files and sim links:

 ~$ ll /usr/share/javascript/jquery/
 drwxr-xr-x 2 root root   4096 Jan 15 20:44 ./
 drwxr-xr-x 5 root root   4096 Jan 15 20:44 ../
 -rw-r--r-- 1 root root 252879 Jul  5  2013 jquery.js
 lrwxrwxrwx 1 root root     13 Jul  5  2013 jquery.lite.js -> jquery.min.js
 lrwxrwxrwx 1 root root      9 Jul  5  2013 jquery.min.js -> jquery.js
 lrwxrwxrwx 1 root root     13 Jul  5  2013 jquery.pack.js -> jquery.min.js
 -rw-r--r-- 1 root root      5 May 23  2012 version.txt

All of the links reference jquery.js but we want to use the minified
jquery.min.js.  The current puppet config is broken for trusty because
it attempts to copy the link instead of the file. This results in a
broken Toggle CI button on Gerrit because that button depends on jquery.
This change creates a jquery.min.js from jquery.js using yui-compressor.

Change-Id: I18056ca528871d9a96218f0919bbd262da005f31