Work-around grubby boot failure on Fedora 20
Something about Fedora 20 and the rax layout of /boot/extlinux.conf
makes grubby fail to correctly set the correct default when the kernel
package is updated.  I found what appears to be the bug (referenced
inline) but there is no clear solution.  However, testing shows the
rawhide version works OK, so we can just install that.

Change-Id: I0880c850dedada893d6e3e7e922c2994ece74930