Expose puppetmaster_server param on openstack_project::puppetmaster
The 'server' value on puppet.conf is currently hardcoded on server.pp,
which is contained within puppetmaster.
This change exposes that parameter in the outer class and in case
the user does not define it then 'puppetmaster.openstack.org' is used
, as is now.
Also added a hiera lookup on the puppetmaster node that again defaults
to puppetmaster.openstack.org. This is to not leave this sensitive
value to a class param default and really put a proper value in hiera.

Change-Id: I8e25397a89cf4adea4decab95de3b5c9ac817d16