Switch to consuming cloud auth from clouds.yaml
Now that nodepool supports creating client objects via shade, we can
consume the auth information (and other associated per-cloud-provider
config) from clouds.yaml. This is an opt-in and worth us doing because
there are other things that can also read the file, which removes
potential duplication of configuring what our clouds are and how to
connect to them.

Depends-On: Iba84b8d578efa8cb7f5af85a5ffbc97f945a47c9
Change-Id: I15008009ccb1b95478a217a81588842742e3a38c