Rename summit server to design-summit-prep
So as to avoid confusion around the summit.openstack.org domain
name, use design-summit-prep.openstack.org for our actual management
name of the server which will allow us to float the summit DNS name
where and when needed. Also indicate that it now runs on trusty.

Once this merges, a new design-summit-prep.openstack.org server will
be launched and the old summit.openstack.org server retired for
eventual deletion.

Change-Id: I410620e2e6ea13b6a6fcb4dee13bd5b16c177a39