Install puppet from the Ubuntu apt repository for aarch64/arm64
The Puppet Labs apt repository does not provide packages for arm64 architecture, preventing
this script from installing puppet on arm64 systems running Ubuntu.  This prevents the use
of Nodepool and the nodepool scripts in openstack-infra/project-config from being used unmodified.

This change determines the host architecture, and, if aarch4/arm64, skips the direct download of
the puppet .deb from apt.puppetlabs.com, instead just installing puppet directly from the Ubuntu
apt repository.

In addition, the Puppet Labs apt repository is not added to the apt configuration if the host is
aarch64/arm64.

Change-Id: Ic4c56d87f7b6b4855a61091e7efcfc2920c4f4bb