Fix smartcn analyzer warnings on puppet-run
Copy to smartcn analyzer to solr lib directory when the target not
exists. Puppet throws this error when the extracted solr
under /tmp/solr- directory gone away after a reboot. The smartcn
analyzer not included in vamsee-solr module, so we needed to
apply this workaround for proper Chinese solr index and search.

Change-Id: Ia19c089af5c3079c98b43b4ce6170f23a0cef3d3