Bump stankevich-python version
Prior versions of stankevich-python have problems on virtualenv
paths. As we install virtualenv from pip, it sets it on
/usr/local/bin. Initially stankevich-python module didn't have
/usr/local/bin on the list of paths to search for virtualenv
binary. It was causing modules using python::virtualenv to
fail, as they could not find the binary.

So we sent a pull request to add /usr/local/bin path to the
list of paths to search virtualenv for:
$path = [ '/bin', '/usr/bin', '/usr/sbin', '/usr/local/bin' ],

Version that brings that is 1.9.4 version, so we need it.

Change-Id: I63b24c92cab9b81ffcde15a36e68815ae08091a0