Add Gemfile and puppet4 checks
In anticipation of upcoming puppet4, it would be nice to check that
we're not going to be broken by it before it happens. There are some
puppet lint plugins to help with that, but in order to use them, we need
to install them.

Rather than installing them globally as we have been doing, copy the
setup used by the puppet-openstack crowd and have a Gemfile so that we
can use bundler.

Also, for now, disable the two most chatty lint checks that we don't
care about in this repo so that we can actually see the errors that will
cause of puppet 4 problems.

NOTE: This is a different Gemfile and Rakefile excludes as are expected
to be put into all of the other puppet repos, because system-config does
not need to be as "clean" as the others in general.

Change-Id: I2828146149fc3527d14fb0a5a9479c4f7314c1ab