Fix puppet lint issues in anticipation of puppet4
In order for us to catch puppet4 issues before they happen, we should
run the puppet4 lint checks. However, if the normal lint checks are too
chatty, we'll never see any puppet4 issues. Clean the ones that aren't
crazy to clean - there are three we'll just suppress because they're
crazy in this repo.

Change-Id: Iface67ce46566aab6bfc0651ace7565d6537a793