Stop installing puppet lint gems globally
ruby has bundler, which is similar to virtualenv. This means there is
no need to install puppet lint gems globally for all nodes, just that
we need to make sure bundler is installed.

NOTE: This should not merge until all of the puppet repos have moved to
using Gemfile. Didn't list that here with depends-on because it's 75
patches and that's probably harder to read.

Change-Id: I87ff9f1eb26bac7c54d3b8f5288a248e81c3c513