Revert "Revert "Turn off HP in nodepool""
This reverts commit 45328d7582e11f4a2019f7fcf1a428a3e3a45b02.

HPCloud isn't actually done cleaning up things on their end and we are
noticing that nodepool is still having trouble with the API there.
Unfortunately this means that we actually have fewer nodes for jobs than
if we just used rax alone. So switch back to rax alone until HPCloud is
properly ready for us.

Change-Id: I274c1f05849b861e9e08d367d55c5bf45a759392