Run more puppet agents at a time with ansible
Our current puppet run_all.sh script takes almost 45 minutes to run
puppet agent on all of our nodes. We are using the default concurrency
of 5. Our puppet master should be able to handle a bit more than that.

Run the git/gerrit playbook with a concurrency of 10 and everything else
with a concurrency of 20.

Change-Id: Ia09abb6fa8c699e156aed38d86ce6fd193f3a42d