Index screen-barbican logs
gate-barbican-devstack-dsvm is a gate job for barbican.

According to elastic-recheck it's been failing pretty hard but without
indexing the barbican logs it's hard to know what's the root failure we
should focus on (and fingerprint in elastic-recheck).

This adds logstash indexing for the screen-barbican log.

Note that the first part of the logs is the output of running
barbican.sh which are not oslofmt so they will likely be squashed
together into a single event, but after that the logs are oslofmt and we
can use them in logstash for fingerprinting failures - and therefore
something > nothing.

Example found here: http://goo.gl/K4d35w

Change-Id: I20f905987a32adac05286c5e6deefe010f541ea2