create proxy for testing hideci
This is a small tests server for testing hideci (and other gerrit
static file changes), it does this by creating a local http server,
then proxying and translating pages on the file (changing hard coded
references to zuul.o.o and review.o.o to route through this
file). You'll end up with a localhost:8001 (port is configurable) that
you can browse to and be on live gerrit + local files served.

Login doesn't work, because the openid redirect magic appears to
happen in a way that we can't intercept (boo), but you can still poke
at a lot of the interface pretty seamlessly this way. Makes for much
better testing.

We need to fetch zuul targeted status urls over http as well because
default python 2.7 apparently blows it's stack on zuul's ssl cert.

Change-Id: Ic28729cea9f2d1aea675025e5fef29f2bcc25164