Start building ubuntu-trusty nodes
As the next step in dib-nodepool, start making nodes called
ubuntu-trusty based on the ubuntu-minimal element which is now in
diskiamge-builder 0.1.44. No jobs should be targetted at this label yet,
however, it is named with the new naming in mind, $distro-$release.

Once these exist, we should be able to start aiming some experimental
devstack jobs at them to make sure that nothing is missing.

Change-Id: I75838d89b6ad7271d8022abbaa0aa4cca3e16525