Add script to split the secret hieradata file
Although this only needs to be done once, it's ugly enough of a task
that it should have a script, and that script should be reviewed.

The script can be run locally on a machine without a hieradata and it
will create a local directory with all of the keys and none of the data.

It will also print out the total mapping and a report on which keys are
duplicated across hosts after group de-duplication is taken in to
account.

Change-Id: Id0af46abb8f4167bd91e95cd7a6cd2e35e5b0a94