WIP: Jenkins slave for package builds
This patch is the first step towards a generic package build environment for
packagers.  Right now, we using Ubuntu 14.04 to install mock and create a
chroot for Fedora 20.

The goal would be to create a new prepare_node_pkg.sh script which nodepool
would use. Creating a new jenkins slave specifically for packaging.

Change-Id: Id33cad905d332cb4a6af810770fb770f385f3207
Signed-off-by: Paul Belanger <pabelanger@redhat.com>