Tell nodepool/shade that rax clouds are rax
In order for shade to know it is talking to rackspace when running in
nodepool we need to set the cloud value in our config to rackspace for
each rackspace provider. This ensures that glance as configured by shade
knows to use the rackspace glance v2 + tasks image upload system.

Change-Id: Ic823a74fffe3a8e51bb8bbaa599fac5f1e731044