Add Fedora 22 DIB nodes
Working on bringing up Fedora 22 nodes.  The dependent chain will need
to be merged to diskimage-builder; still verifying nodepool element &
devstack-gate scripts but having the logs from real builds will be
helpful.

Note the plan is get rid of F21 nodes as soon as this is ready and
doing devstack CI.

Change-Id: I4ca94054b6f96f6e3b258517bdf12ce070692ffa
Depends-On: Icbd08fb5aa69446ad65ff72af631902c4e1fa12b
Depends-On: Ic8b0fd2a19cbb90e1d57e3025a3d2513ebada2e2
Depends-On: Iadb41317a1ad217a9b90c015d22a09ee214c4594
Signed-off-by: Paul Belanger <pabelanger@redhat.com>