Add Fedora 22 DIB nodes
Working on bringing up Fedora 22 nodes.  The dependent chain will need
to be merged to diskimage-builder; still verifying nodepool element &
devstack-gate scripts but having the logs from real builds will be
helpful.

Note the plan is get rid of F21 nodes as soon as this is ready and
doing devstack CI.

Change-Id: I4ca94054b6f96f6e3b258517bdf12ce070692ffa
Depends-On: Icbd08fb5aa69446ad65ff72af631902c4e1fa12b
Signed-off-by: Paul Belanger <pabelanger@redhat.com>