Start using jfryman/puppet-selinux
We were using a fork of that original module, called
spiette/puppet-selinux. This module had no maintenance since
2013, and it's failing under puppet 3.

We have found this issue:
http://elatov.github.io/2014/08/messing-around-with-puppet/

So far, there are pull requests into that module since 2014,
and there is no activity on that. The original jfryman module
is up to date, doesn't have this problem, and is actively
maintaned. So start using that and deprecate the spiette one,
in order to be able to use selinux again.

Change-Id: I053d1437ef85828654a2a3bb9fedd6296525dabd
Depends-On: I8500da48191613d2fc7451bc829c27e9e618711e