Add libuuid devel packages
python-qpid-proton does something similar to pyzmq when libzmq is not
installed. It bundles the underlying library and builds it.
Unfortunately, python-qpid-proton currently has a dependency on
`libuuid` and therefore, it depends on the devel packages to be present.

This dependency will be removed in the future but it'll take time. Until
that happens, I'm adding uuid-dev to the list of packages to install by
default.

Change-Id: Iae4d7138311eec0cd90c0865f30b84a98ce3c142