Change logstash multiline stream identity
Change the multiline stream identity to be based off of the nodepool
node rather than the host the logstash indexer is running on. An example
host value seen in kibana is '127.0.0.1: 60499' and refers to tcp source
used by the gearman worker when sending data to the logstash tcp input.
While it does work to use host because it is unique to the gearman
worker and a gearman worker will only be pushing one log at a time, it
makes more sense to uniquely identify the stream based on the log file
source rather than the gearman worker pushing the file.

Change-Id: Iad2087000ee8e4eec4a8d71c96f9e585cb4595ea