Change logstash multiline stream identity
It is confusing to base the multiline stream identity on both the host that
the indexer process is running on as well as the filename being indexed,
but not on the host the job ran on. But this was needed due to a bug in
logstash 1.3.3 that required an event that does not match the multiline
rule to trigger a flush of the multiline event.

Newer versions of logstash are able to regularly flush the multiline
event buffer (default after 5 seconds), so once we upgrade to logstash
2.0 we can base the stream identity off of zuul's build uuid combined
with the filename.

Change-Id: Iad2087000ee8e4eec4a8d71c96f9e585cb4595ea