Load in nanliu/git module
Vcsrepo has some problems, largely due to the 'be everything for
everyone' attitude. We have our own untested an unmaintained fork.
What we really want is a well written type and provider for git.
That is this module.

No License is yet attached to the repo, I've created an Issue for that.
I've also submitted a PR to add tag support.

Change-Id: I91b7e35ef5e66c3538fd987d3857ec1af455beb1