Fix yaml2ical symlinks
Because the scp publisher insists on keeping the source folder
from the workspace (when uploading with keep-hierarchy which
is necessary to get all of the subdirs) we should update where
the symlinks point.

The DocumentRoot for eavesdrop needs a symlink to the calendars dir
now that yaml2ical places individual icals in there.

For this to work we need to wait for the output dir to be created
from a successful run on the publisher job.

Change-Id: Icd6fd2969fcea3ef9b7ecd7d482941dbd92b2d26
Depends-On: Ic968a1883d3942335547ed52948030f7f1f84add