WIP - reprepro UCA mirror and corresponding override/overlay area
This is a PoC meant to inspire discussion on
https://review.openstack.org/#/c/179713/
(Ic7b701f3882382eb67edadfa3093a8a8034d7bee)
and in response to https://storyboard.openstack.org/#!/story/2000175

If the relevant questions are what to build, how to build, and
how to publish, this changeset is one potential answer to only
the last question.

To that end, it uses reprepro to mirror UCA's "trusty-updates/kilo",
inoticoming to trigger reprepro when packages are uploaded, placing
them then into "trusty-kilo/overlay", and an unsophisticated Apache
config to serve up the trusty-updates/kilo mirror and
trusty-kilo/overlay.

Uploading can presently be achieved through scp to
debupload@debrepo.openstack.org:/srv/debrepo/incoming .

Change-Id: I84b125191be47c64aa7258f5c82b5671a78090b4
Depends-On: Ic7b701f3882382eb67edadfa3093a8a8034d7bee