Add cfacter to puppet3 nodes
CFacter is a rewrite of facter in C. It can live next to ruby facter and
will replace facter when puppet.conf is set to do so. For now, put it on
machines so we can get used to it and experiment with using it speed
things up. The primary advantage is that it is faster.

Change-Id: I3433a670eaf7718cb20311835b9109504312a2ef