Added NPM Publishing credentials to pypi slave.
This patch adds an .npmrc file for the jenkins user on our pypi slave.
Note that actually publishing to npm will only work on trusty or later.
This patch requires a manual step, where a member of infra-root registers
the openstackci user on npm, and injects those credentials into hiera.
Note that npm credentials require a user, password, and URL.

Change-Id: I13fa91500a7b189e5fb6e915968727e8c1f2a1f4