Trim the elements list from nodepool.yaml
We are listing a whole lot of elements, but elements have depenendy
information inside of them. Instead of listing all of them, we should
list the key ones that provide the things we need and let their
dependency graph sort out the rest. This way we can refactor the
elements themselves without having to do a dance with the nodepool
config every time.

Change-Id: Id6494b7c88b652f4d6fb4777e8f6e4005dc56321
Depends-On: Ib2296bed1d7f289e7aa2720a38555275de1b6402