Ensure a symlink for pip-python
Puppet on redhat expects pip to be avilable as pip-python and doesn't
check for a 'pip' as a fallback. Work around this bug by creating a
symlink. Usually we escape needing to do this because install_puppet.sh
does this work. Adding to puppet so that all the servers stay in sync.

Change-Id: I1143d19cade20609bbcc4c1504704ff2c45dfdcd