(WIP) Create playbook to puppet InfraCloud nodes
We don't have a masterless puppet environment set up yet, but we should
start out Infra Cloud masterless anyway. This adds a task to install
the OpenStack Puppet Modules using infracloud-modules.env, copy local
hiera data into place, and run puppet apply with the run_puppet.sh
script.

The hiera hierarchy that is copied onto the node is based on the
hierarchy documented in
system-config/modules/openstack_project/files/puppetmaster/hiera.yaml.
We only copy the data file if it exists on the ansible master, because
not all hosts will use public hiera or have their own fqdn or group
hiera data files. The exception is the secret common.yaml, which we
know must exist, so we copy it unconditionally. This should be reusable
for when other nodes move to a masterless setup.

Since run_puppet.sh logs puppet output to a regular file, this patch
adds a logrotate rule for that logfile to the openstack_project::server
class. This means the logrotate rule will be applied to all nodes, but
since the options contain 'missingok' it will not have an impact on
nodes still running in a master-agent setup. The logrotate class is not
explicitly included because it is already included within the
logrotate::file defined type.

todo:

 - puppetdb connection

This should probably eventually become an ansible role instead of a
task.

Change-Id: I8226fb338fad4d3df11957a8710c4a9d5984048d