Keep npm up-to-date.
This patch makes sure the version of NPM is up-to-date
with the most recent one. Note that this will not update
nodejs, so the "latest" version of NPM is actually restricted
to the one that will run on the installed version of node.

Furthermore, the upgrade script has been restricted to run
once daily. Calling out to npm every 15 minutes for a cache
update is unecessary.

Change-Id: Icce880e7f1d0df427e6cf55db6e8cd98e816c141