Disable DIB internal APT cache for nodepool
The default behavior of the dpkg DIB element is to overlay a
directory from the host filesystem at /var/cache/apt/archives in the
chroot so that the cache of packages can persist to subsequent image
builds. This has a side effect of not actually populating the cache
within the chroot at all. Disable the local cache on the host so
that we can cache DEB packages in our images.

Change-Id: I1e33cb2e6049c83b3aad4a18885d8bd909f09195