Collect ssh job test results in the DB
The gate-tempest-dsvm-neutron-full-ssh job is non-voting and
running on check only, until we can verify whether it's stable
enough to be promoted as voting.

We are also in the process of enabling ssh verification in more
tests, and we want monitor whether these changes introduce any
degration.

To help in this we would like to enhance the data available
from statsd with test level data from the subunit2sql DB.

Partially-implements bp ssh-auth-strategy

Change-Id: Iead11b35e29792a66d540b6b331d3e160248227c